feat: Added The Auto Updater Plugins and Modified the file structure for the components. (#33)
* fix: Added Auto Updater Plugins

* feat: Modified core components for schema exploration, query building, SQL workspace, and ER diagram visualization.

* fixes the header placement

* feat: Add footer Schema Explorer panel and bumping the version.
