+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "relwave",
   "private": true,
-  "version": "0.1.0-beta.2",
+  "version": "0.1.0-beta.3",
   "type": "module",
   "scripts": {
     "dev": "vite",
