+++ src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "RelWave",
-  "version": "0.1.0-beta.2",
+  "version": "0.1.0-beta.3",
   "identifier": "com.yashs.RelWave",
   "build": {
     "beforeDevCommand": "npm run dev",
