+++ src/components/database/index.ts
@@ -1,2 +1,1 @@
-export { default as SqlEditor } from './SqlEditor'
 export { default as MigrationsPanel } from './MigrationsPanel'
