+++ src/components/query-builder/index.ts
@@ -1,3 +1,23 @@
-export { default as ControlPanel } from './ControlPanel'
-export { default as QueryBuilderHeader } from './QueryBuilderHeader'
-export { default as VisualBuilder } from './VisualBuilder'
+// Main panel component
+export { default as QueryBuilderPanel } from './QueryBuilderPanel';
+
+// Sub-components
+export { BuilderHeader } from './BuilderHeader';
+export { BuilderSidebar } from './BuilderSidebar';
+export { DiagramCanvas } from './DiagramCanvas';
+export { SQLResultsPanel } from './SQLResultsPanel';
+export { BuilderStatusBar } from './BuilderStatusBar';
+
+// Types
+export type {
+    QueryFilter,
+    ColumnOption,
+    QueryHistoryItem,
+    TableSchema,
+    QueryProgress,
+    BuilderHeaderProps,
+    BuilderSidebarProps,
+    DiagramCanvasProps,
+    SQLResultsPanelProps,
+    BuilderStatusBarProps,
+} from './types';
