+++ src/components/schema-explorer/metadata/index.ts
@@ -1,0 +1,5 @@
+export { DatabaseDetails } from './DatabaseDetails';
+export { SchemaDetails } from './SchemaDetails';
+export { TableDetails } from './TableDetails';
+export { ColumnDetails } from './ColumnDetails';
+export { EnumDetails, SequenceDetails } from './OtherDetails';
