+++ src/components/workspace/index.ts
@@ -1,0 +1,12 @@
+// Main panel component
+export { default as SQLWorkspacePanel } from './SQLWorkspacePanel';
+
+// Sub-components
+export { WorkspaceHeader } from './WorkspaceHeader';
+export { WorkspaceSidebar } from './WorkspaceSidebar';
+export { QueryTabBar } from './QueryTabBar';
+export { ResultsPanel } from './ResultsPanel';
+export { StatusBar } from './StatusBar';
+
+// Types
+export type { QueryTab, QueryHistoryItem, TableInfo } from './types';
